deps: update v8 to 4.4.63.9
Upgrade the bundled V8 and update code in src/ and lib/ to the new API.

Notable backwards incompatible changes are the removal of the smalloc
module and dropped support for CESU-8 decoding.  CESU-8 support can be
brought back if necessary by doing UTF-8 decoding ourselves.

This commit includes https://codereview.chromium.org/1192973004 to fix
a build error on python 2.6 systems.  The original commit log follows:

    Use optparse in js2c.py for python compatibility

    Without this change, V8 won't build on RHEL/CentOS 6 because the
    distro python is too old to know about the argparse module.

PR-URL: https://github.com/nodejs/io.js/pull/2022

+++ deps/v8/src/startup-data-util.cc
@@ -44,7 +44,7 @@ char* StartupDataHandler::RelativePath(char** buffer, const char* exec_path,
   DCHECK(exec_path);
   const char* last_slash = strrchr(exec_path, '/');
   if (last_slash) {
-    int after_slash = last_slash - exec_path + 1;
+    int after_slash = static_cast<int>(last_slash - exec_path + 1);
     int name_length = static_cast<int>(strlen(name));
     *buffer = reinterpret_cast<char*>(calloc(after_slash + name_length + 1, 1));
     strncpy(*buffer, exec_path, after_slash);
@@ -75,7 +75,7 @@ void StartupDataHandler::Load(const char* blob_file,
   if (!file) return;
 
   fseek(file, 0, SEEK_END);
-  startup_data->raw_size = ftell(file);
+  startup_data->raw_size = static_cast<int>(ftell(file));
   rewind(file);
 
   startup_data->data = new char[startup_data->raw_size];
